+++ src/core/mod.rs
@@ -6,11 +6,13 @@
 mod macros;
 
 mod arc;
+mod four_char_code;
 mod object_type;
 mod os_err;
 mod os_status;
 
 pub use arc::*;
+pub use four_char_code::*;
 pub use object_type::*;
 pub use os_err::*;
 pub use os_status::*;